+++ spec/features/journal_administration_spec.rb
@@ -12,7 +12,6 @@ feature "Journal Administration", js: true do
 
   let(:admin_page) { AdminDashboardPage.visit }
   let(:journal_page) { admin_page.visit_journal(journal) }
-  let(:flow_page) { journal_page.add_flow }
 
   describe "journal listing" do
     context "when the user is a site admin" do
@@ -65,54 +64,5 @@ feature "Journal Administration", js: true do
         expect(journal_page).to have_no_mmt_name(mmt_to_delete.paper_type)
       end
     end
-
-    describe "on a Journal's Flow Manager" do
-
-      describe do
-        before { flow_page }
-
-        it "show Journal name as text" do
-          expect(page.find(".column-title-wrapper")).to have_content journal.name
-        end
-      end
-
-      describe do
-        it "show Journal logo" do
-          using_wait_time 5 do
-            with_aws_cassette(:yeti_image) do
-              journal.update_attributes(logo: File.open("spec/fixtures/yeti.jpg"))
-              visit "/admin/journals/1/old_roles/1/flow_manager"
-              find(".control-bar-link-icon").click
-              expect(page.find(".column-title-wrapper")).to have_css("img")
-            end
-          end
-        end
-      end
-
-      context "editing a flow title" do
-        before { flow_page }
-
-        it "reveals 'cancel / save' buttons" do
-          expect(page).to have_content 'Up for grabs'
-          find('h2.column-title').click()
-          expect(page).to have_button('cancel')
-          expect(page).to have_button('Save')
-        end
-      end
-
-      context "saving an edited flow" do
-        before { flow_page }
-
-        it "removes 'cancel / save' buttons" do
-          expect(page).to have_content 'Up for grabs'
-          find('h2.column-title').click()
-          find('button.column-header-update-save').click()
-          expect(page).not_to have_button('cancel')
-          expect(page).not_to have_button('Save')
-        end
-      end
-
-    end
-
   end
 end