+++ client/tests/addons/add-author-test.js
@@ -116,7 +116,6 @@ module('Integration: adding an author', {
     ]);
 
     server.respondWith('GET', '/api/admin/journals/authorization', [204, { 'Content-Type': 'application/json' }, '' ]);
-    server.respondWith('GET', '/api/user_flows/authorization', [204, { 'Content-Type': 'application/json' }, '' ]);
     server.respondWith('GET', '/api/affiliations', [200, { 'Content-Type': 'application/json' }, JSON.stringify([]) ]);
     server.respondWith('GET', '/api/journals', [
       200, {