+++ spec/factories/user_flow_factory.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :user_flow do
-    user
-  end
-end