+++ engines/tahi_standard_tasks/client/app/models/funder.js
@@ -25,16 +25,19 @@ export default NestedQuestionOwner.extend({
 
   funderHadInfluence: Ember.computed.alias('funderInfluenceAnswer.value'),
 
-  funderInfluenceAnswer: Ember.computed(function() {
+  funderInfluenceAnswer: Ember.computed('nestedQuestionAnswers.[]', function() {
     return this.answerForQuestion('funder--had_influence');
   }),
 
   funderInfluenceDescription: Ember.computed.alias(
       'funderInfluenceDescriptionAnswer.value'),
 
-  funderInfluenceDescriptionAnswer: Ember.computed(function() {
-    return this.answerForQuestion('funder--had_influence--role_description');
-  }),
+  funderInfluenceDescriptionAnswer: Ember.computed(
+      'nestedQuestionAnswers.[]',
+      function() {
+        return this.answerForQuestion('funder--had_influence--role_description');
+      }
+  ),
 
   onlyHasAdditionalComments: Ember.computed(
       'additionalComments',