+++ CHANGELOG.md
@@ -5,6 +5,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ## [{version}] - {release_date}
 ### Added
 - Funders now have an "additional comments" field
+- Email report to admin counting papers in states
 
 ### Changed
 -